debug - update unzip directory, delete unziped files

+++ pvr/GuiHelper.py
@@ -417,19 +417,49 @@ def RemoveDirectory( aPath ) :
 	return ret
 
 
-def RemoveUnzipFiles( aUsbPath, aUnzipFile ) :
-	if not CheckDirectory( aUnzipFile ) :
-		LOG_TRACE('not exist unzipfile[%s]'% aUnzipFile )
-		return
+def RemoveUnzipFiles( aUsbPath, aZipFile = False, aReqFile = False ) :
+	fileList = None
 
-	fileList = GetUnpackFiles( aZipFile )
-	if not fileList :
-		return False
+	try :
+		if aZipFile :
+			if not CheckDirectory( aZipFile ) :
+				#LOG_TRACE('not exist unzipfile[%s]'% aZipFile )
+				return False
+			fileList = GetUnpackFiles( aZipFile )
 
-	for iFile in fileList :
-		delFile = '%s/%s'% ( aUsbPath, iFile[0] )
-		RemoveDirectory( delFile )
-		LOG_TRACE('delete file[%s]'% iFile[0] )
+			if not fileList or len( fileList ) < 1 :
+				#LOG_TRACE( 'no delete files' )
+				return False
+
+			for iFile in fileList :
+				if iFile[0].strip( ) :
+					delFile = '%s/%s'% ( aUsbPath, iFile[0].strip( ) )
+					RemoveDirectory( delFile )
+					#LOG_TRACE('delete file[%s]'% delFile )
+
+		else :
+			if not CheckDirectory( aReqFile ) :
+				#LOG_TRACE('not exist reqFile[%s]'% aReqFile )
+				return False
+
+			fd = open( aReqFile, 'r' )
+			fileList = fd.readlines( )
+			fd.close( )
+
+			if not fileList or len( fileList ) < 1 :
+				#LOG_TRACE( 'no delete files' )
+				return False
+
+			for iFile in fileList :
+				if iFile.strip( ) :
+					delFile = '%s/%s'% ( aUsbPath, iFile.strip( ) )
+					RemoveDirectory( delFile )
+					#LOG_TRACE('delete file[%s]'% delFile )
+
+	except Exception, e :
+		LOG_ERR( 'except[%s]'% e )
+
+	return True
 
 
 def CheckDirectory( aPath ) :
@@ -616,7 +646,8 @@ def UnpackToUSB( aZipFile, aUsbPath, aUnpackSize = 0, aUnzipPath = 'update_ruby'
 		LOG_TRACE( '------------check usb[%s]'% aUsbPath )
 		return isCopy
 
-	RemoveDirectory( '%s/%s'% ( aUsbPath, aUnzipPath ) )
+	#RemoveDirectory( '%s/%s'% ( aUsbPath, aUnzipPath ) )
+	RemoveUnzipFiles( aUsbPath, False, aUnzipPath )
 	usbSize = GetDeviceSize( aUsbPath )
 	if usbSize <= aUnpackSize :
 		return -1
